refactor confBaudRate() to correct some API inconsistency
confBaudRate() in some cases didn't clearly return a boolean, though it
is supposed to.  The code also lacked clarity, having several
ambiguities.  This change is more significant than is strictly
necessary, but so many changes were advisable that it was easier to
largely rewrite it. There is now one exit point with a clear return
value which now is clearly set properly at different possible points in
the function. Made some readability improvements too.